Refactory of users objects, done for now with Authen/Author
