--- src/AppBundle/Entity/User/Person.php
@@ -1,125 +1,0 @@
-<?php
-
-namespace AppBundle\Entity\User;
-
-use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Security\Core\User\UserInterface;
-
-/**
- * 
- */
-abstract class Person
-{
-    /**
-     * @var int
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    private $id;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="firstName", type="string", length=70)
-     */
-    private $firstName;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="lastName", type="string", length=155)
-     */
-    private $lastName;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="mail_address", type="string", length=255, unique=true)
-     */
-    private $mailAddress;
-
-
-    /**
-     * Get id.
-     *
-     * @return int
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * Set firstName.
-     *
-     * @param string $firstName
-     *
-     * @return User
-     */
-    public function setFirstName($firstName)
-    {
-        $this->firstName = $firstName;
-
-        return $this;
-    }
-
-    /**
-     * Get firstName.
-     *
-     * @return string
-     */
-    public function getFirstName()
-    {
-        return $this->firstName;
-    }
-
-    /**
-     * Set lastName.
-     *
-     * @param string $lastName
-     *
-     * @return User
-     */
-    public function setLastName($lastName)
-    {
-        $this->lastName = $lastName;
-
-        return $this;
-    }
-
-    /**
-     * Get lastName.
-     *
-     * @return string
-     */
-    public function getLastName()
-    {
-        return $this->lastName;
-    }
-
-    /**
-     * Set mailAddress.
-     *
-     * @param string $mailAddress
-     *
-     * @return User
-     */
-    public function setMailAddress($mailAddress)
-    {
-        $this->mailAddress = $mailAddress;
-
-        return $this;
-    }
-
-    /**
-     * Get mailAddress.
-     *
-     * @return string
-     */
-    public function getMailAddress()
-    {
-        return $this->mailAddress;
-    }
-}
